+++ R/fnb.update.R
@@ -145,6 +145,30 @@ fnb.update.fnb.gaussian <- function(object, x, y, sparse = FALSE, check = TRUE){
 #' @export
 #' @import Matrix
 #' @rdname updateFNB
+fnb.update.fnb.poisson <- function(object, x, y, sparse = FALSE, check = TRUE){
+  if(check){
+    args <- fnb.check.args.model(x, y, priors=NULL, sparse)
+    x <- args$x
+    y <- args$y
+    sparse <- args$sparse
+  }
+
+  oldx <- object$x
+  x <- x[,colnames(x),drop=FALSE]
+  if(ncol(x)!=ncol(oldx)){
+    stop("x has different columns than original data used to build object")
+  }
+
+  newx <- rbind(object$x, x)
+  newy <- factor(c(as.character(object$y), as.character(y)))
+
+  return(fnb.poisson(newx, newy, object$priors, sparse))
+}
+
+
+#' @export
+#' @import Matrix
+#' @rdname updateFNB
 fnb.update.fastNaiveBayes <- function(object, x, y, sparse = FALSE, check = TRUE){
   if(check){
     args <- fnb.check.args.model(x, y, priors=NULL, sparse)